chore(types): resolve 57 mypy errors across 15 files → 0
- cli.py: add TYPE_CHECKING guard for Database / ResourceInventory so
  the string annotations resolve (previously `Name not defined`).
- database.py / inventory.py: Iterator[sqlite3.Connection] on the
  contextmanager-decorated get_connection; assert lastrowid before
  returning int.
- parser.py: annotate `actions` / `resources` local lists.
- analyzer.py: widen get_approval_stats return to Dict[str, float].
- hmac_keys.py: accept Mapping[str, object] for sign_row/verify_row
  (invariance workaround for dict[str, str] callers).
- net/guards.py: str() narrow on getaddrinfo sockaddr[0].
- cli.py refresh: split ManagedPoliciesLoader / CloudSplainingLoader
  into distinct variables (had one Union-typed name causing
  incompatible-assignment + missing-attribute errors).
- cli_cache._human_bytes: use float accumulator.
- logging_setup: narrow type:ignore comments on structlog processor
  protocol mismatches.
- Removed 5 stale `# type: ignore` comments flagged by warn_unused_ignores.

mypy src/sentinel --explicit-package-bases: 0 errors.

+++ src/sentinel/analyzer.py
@@ -1165,7 +1165,7 @@ class HITLSystem:
         """
         return self.decisions.copy()
 
-    def get_approval_stats(self) -> Dict[str, int]:
+    def get_approval_stats(self) -> Dict[str, float]:
         """Get approval statistics.
 
         Returns:
